app/api/requirements.txt
flask
requests
grpcio
google-cloud-storage
google-cloud-bigquery
google-cloud-automl
google-cloud-logging
pyjwt
flasgger
waitress